requirements.txt
tensorflow-gpu==2.4.0
pandas==1.3.3
sklearn==0.0